JPP BZ 966890 - Neutral product names in i18n.
Changed a few messages in the resources files for several languages to make it more product neutral. A review for the following languages would be good, to confirm that the meaning didn't get compromised:
- French
- Dutch
- Russian
- Ukrainian
- Vietnamese

Also, fixed typos and some expressions, as well as missing translations, in the Brazilian Portuguese resources file.