GTNPORTAL-3379 - Added a Clear Search button to some search pages
On some search pages, there's no way to get back to the initial state of the listing without changing pages. From a user experience perspective, this is not desirable. This commit introduces a new icon at the right side of the magnifying glass (ie: search) to clear the search restrictions, effectively returning the data table to the initial state.