GTNPORTAL-3402 - Custom password policies
Added the capability of setting custom password policies. If the configuration entry "passwordpolicy" is found on the configuration, it's then used for the password validation. Otherwise, it fallsback to the previous behavior, which is to validate that a password has a minimum of 6 chars and maximum of 30 chars.

Added a method on the UserConfigurableValidator that lists the currently known configuration entries for this validator, and a test for that.