GTNPORTAL-3432 - Replaced all JBoss-related profiles with eap630.
As part of GTNPORTAL-3432, this commit removes all JBoss-related profiles, including previous EAP versions, with a single profile for EAP 6.3.0.Alpha1. Part of this change is the rename of the packaging/jboss-as7 into jboss.

The generic 'jboss' name was chosen as it would be an indicative of the 'base' of the integration, meaning that it would still be correct when Wildfly gets integrated.